backend/requirements.txt
fastapi
requests
uvicorn
openai
python-dotenv
python-multipart
beautifulsoup4
python-jose[cryptography] 
passlib[bcrypt] 
sqlalchemy
boto3